✨ feat(RegenerateEncryptData.php): add new action to regenerate encrypted data in the database
The `RegenerateEncryptData` class is a new action that is responsible for regenerating encrypted data in the database. This action is invoked to update the encrypted data using a new crypt key and iv.

The action performs the following steps:
1. Generates a new crypt key and iv using `Str::random()` method.
2. Begins a database transaction.
3. Iterates over all the files in the `app/Models` directory.
4. For each model, it retrieves the casts and identifies the columns that are encrypted using the `Encrypted` cast.
5. If there are encrypted columns, it calls the `reEncryptData()` method to re-encrypt the data using the new crypt key and iv.
6. Commits the database transaction.
7. Calls the `replaceCryptKeyAndIv()` method to replace the old crypt key and iv with the new ones in the `.env` file.

The `replaceCryptKeyAndIv()` method replaces the old crypt key and iv with the new ones in the `.env` file. If the `.env` file already contains `CRYPT_PREVIOUS_KEY` and `CRYPT_PREVIOUS_IV`, it updates the values. Otherwise, it adds the new crypt key and iv along with the previous key and iv.

The `reEncryptData()` method re-encrypts the data in the database for a given model and encrypted columns. It retrieves the models in chunks, locks them for update, and iterates over each model. For each encrypted column, it decrypts the data using the old crypt key and iv, and then encrypts it using the new crypt key and iv. Finally, it updates the model with the new encrypted data.

This action is useful when the crypt key and iv need to be rotated to enhance security or when the encryption algorithm or configuration has changed.