🔧 chore(RegenerateEncryptedDataCommand.php): add new command to regenerate encrypted data in the database 🔧 chore(StarterKitServiceProvider.php): register RegenerateEncryptedDataCommand in the service provider 🔧 chore(config/crypt.php): add 'regenerate' option to enable or disable regeneration of encryption key and iv
The changes were made to add a new command called RegenerateEncryptedDataCommand. This command allows the user to regenerate encrypted data in the database. The command prompts the user for the encryption key in the .env file and confirms if they want to regenerate the encrypted data. If the user confirms, the command generates new encryption key and iv, and re-encrypts the data in the database using the new key and iv. The old key and iv are replaced with the new key and iv in the .env file. The command also checks for the 'regenerate' option in the crypt configuration file to determine if the encryption key and iv can be regenerated or not. The RegenerateEncryptedDataCommand is registered in the StarterKitServiceProvider to make it available as a console command.